move db client to separate folder
