+++ scripts/database/seed.ts
@@ -1,6 +1,6 @@
 import { blobToFile } from "@/image/blob";
 import { productNameLengthMax, productNameLengthMin, productRatingMax } from "@/product/validation";
-import { db } from "@/server/database";
+import { db } from "@/server/database/client";
 import type { ReviewInsert } from "@/server/database/schema/product";
 import { category, review, reviewsToCategories } from "@/server/database/schema/product";
 import { user } from "@/server/database/schema/user";
