+++ src/server/api/user/review/delete.ts
@@ -1,5 +1,5 @@
 import { protectedProcedure } from "@/server/api/trpc";
-import { db } from "@/server/database";
+import { db } from "@/server/database/client";
 import { nonNullableSQL } from "@/server/database/query";
 import { review } from "@/server/database/schema/product";
 import { throwExpectedError } from "@/server/error/trpc";
