+++ src/server/api/user/review/upsert.ts
@@ -8,7 +8,7 @@ import {
   productRatingMax,
 } from "@/product/validation";
 import { protectedProcedure } from "@/server/api/trpc";
-import { db } from "@/server/database";
+import { db } from "@/server/database/client";
 import type { ReviewInsert } from "@/server/database/schema/product";
 import { category, review, reviewsToCategories } from "@/server/database/schema/product";
 import { throwExpectedError } from "@/server/error/trpc";
